Reimplemented unquote; Eden: tweaks to draw and find.
When a quote is drawn, the qweight of the drawn quote is halved. This was already the case. Now after this happens, we normalize all qweight values such that their sum is close to 1. Find and remove now use much neater methods of retrieving target quotes. There is some code overlap between the two, and find is called before remove when the unquote interaction is used. Some pruning can be done.