Refactored Eden connection handling. Add timestamps to debug msgs.
Switched from using Tokio to a simpler method using std::net. It is blocking instead of async, but that doesn't seem to be a limitation in this case, but it can be made async if needed. Connections now stay open only as long as needed. The problem with Tokio was that connections simply wouldn't die, causing a crash.